requirements.txt
requests
python-dotenv
pandas
msgraph-sdk
openpyxl
streamlit